refactorizacion y estructuracion
