+++ app/routers/weather.py
@@ -1,19 +1,11 @@
-from fastapi import FastAPI
-from supabase_getter import get_locations_codes
-from weather_fetcher import fetch_weather_forecast, fetch_weather_forecast_by_city_id
-from supabase_upper import save_forecast_to_db, check_locality_exists
+from fastapi import APIRouter
+from app.services.supabase_getter import get_locations_codes
+from app.services.weather_fetcher import fetch_weather_forecast_by_city_id
+from app.services.supabase_upper import save_forecast_to_db
 
-app = FastAPI()
+router = APIRouter(prefix="", tags=["Clima"])
 
-# Ubicación por defecto (puedes ampliarlo a varias)
-DEFAULT_LAT = -2.9
-DEFAULT_LON = -78.99
-
-@app.get("/")
-def root():
-    return {"mensaje": "API del planificador climático activa"}
-
-@app.post("/actualizar_clima/")
+@router.post("/actualizar_clima/")
 def actualizar_clima():
     try:
         locations = get_locations_codes()
@@ -29,15 +21,10 @@ def actualizar_clima():
     except Exception as e:
         return {"status": "ERROR", "detalle": str(e)}
 
-
-
-@app.get("/clima_pronostico/")
+@router.get("/clima_pronostico/")
 def clima_pronostico(city_id: int = 3654157):
     try:
         datos = fetch_weather_forecast_by_city_id(city_id)
         return {"status": "OK", "datos": datos}
     except Exception as e:
         return {"status": "ERROR", "detalle": str(e)}
-    
-
-
