+++ app/services/supabase_upper.py
@@ -1,21 +1,9 @@
-import requests
-import os
 from datetime import datetime
 from dotenv import load_dotenv
-from supabase_client import supabase
+from app.db.supabase_client import supabase
+from app.services.supabase_getter import check_code_exists, check_locality_exists
 
 load_dotenv()
-
-def check_locality_exists(supabase, location_id: int) -> bool:
-    try:
-        result = supabase.rpc("check_locality_exists", {"locality_id": location_id}).execute()
-        # El valor booleano estará en result.data, por ejemplo: [True]
-        if result.data:
-            return True
-        return False
-    except Exception as e:
-        print("Error al verificar existencia de localidad:", e)
-        return False
 
 def save_forecast_to_db(data):
     
@@ -99,16 +87,4 @@ def save_forecast_to_db(data):
 
 
 
-def check_code_exists(supabase, function, code_name, code: int) -> bool:
-    try:
-        result = supabase.rpc(function, {code_name: code}).execute()
-        # El valor booleano estará en result.data, por ejemplo: [True]
-        if result.data:
-            return True
-        return False
-    except Exception as e:
-        print(f"Error al verificar existencia de código {code} en la funcion {function}:", e)
-        return False
-
-
