+++ app/services/weather_fetcher.py
@@ -2,7 +2,7 @@ import requests
 import os
 from datetime import datetime
 from dotenv import load_dotenv
-from supabase_client import supabase
+from app.db.supabase_client import supabase
 
 load_dotenv()
